Structural, price changes
Moved the actual item inclusion out to a library in case I (or someone else)
wants to add a 'placed' option for the items where they can be found in chests,
stores, enemies instead of all in Conlan's store.

Per Wes' to-do list and based on Discord feedback, an initial pass of price
adjustments to the items was made. These were also farmed out to an inactive
library to allow players to make their own adjustments as desired.